working on proper esm support
